requirements.txt
numpy==1.20.2
opencv_python==4.5.1.48
pandas==1.3.5
Pillow==9.5.0
selenium==4.9.1
